#5 evaluation of student wishes and event assignment

+++ src/main/java/itf221/gvi/boom/io/reader/XlsxReader.java
@@ -52,7 +52,7 @@ public class XlsxReader implements Reader {
                         cellList.add(cell.getStringCellValue());
                         break;
                     case NUMERIC:
-                        cellList.add(String.valueOf(cell.getNumericCellValue()));
+                        cellList.add(String.valueOf((int)cell.getNumericCellValue()));
                         break;
                     default:
                         cellList.add("");
